Removed header
Single source file module
