+++ mod_mrf.vcxproj
@@ -22,9 +22,6 @@
     <ClCompile Include="src\mod_mrf.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="src\mod_mrf.h" />
-  </ItemGroup>
-  <ItemGroup>
     <None Include=".gitignore" />
     <None Include="apr_FOPEN_RANDOM.patch" />
     <None Include="LICENSE" />
